+++ src/main/scala/stellar/sdk/inet/HorizonAccess.scala
@@ -66,7 +66,7 @@ class Horizon(uri: URI)
   implicit val formats = Serialization.formats(NoTypeHints) + AccountRespDeserializer + DataValueRespDeserializer +
     LedgerRespDeserializer + TransactedOperationDeserializer + OrderBookDeserializer +
     TransactionPostResponseDeserializer + TransactionHistoryDeserializer + FeeStatsRespDeserializer +
-    NetworkInfoDeserializer
+    NetworkInfoDeserializer + PaymentPathDeserializer
 
   override def post(txn: SignedTransaction)(implicit ec: ExecutionContext): Future[TransactionPostResponse] = {
     for {